pygame/requirements.txt
PyYAML<6.1.0
pygame-ce<2.5.0